add tasks file & handler errors in main
